server/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.119.0
h11==0.16.0
idna==3.11
pika==1.3.2
pydantic==2.12.2
pydantic_core==2.41.4
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
sqlalchemy==2.0.23
python-multipart==0.0.6
rpi-lgpio==0.5  
psutil==5.9.6
